Moved back xcode post copy script to use cp since rsync not working

+++ gameplay/src/Model.cpp
@@ -310,7 +310,7 @@ static bool drawWireframe(MeshPart* part)
     {
     case Mesh::TRIANGLES:
         {
-            for (unsigned int i = 0; i < indexCount; i += 3)
+            for (size_t i = 0; i < indexCount; i += 3)
             {
                 GL_ASSERT( glDrawElements(GL_LINE_LOOP, 3, part->getIndexFormat(), ((const GLvoid*)(i*indexSize))) );
             }
@@ -319,7 +319,7 @@ static bool drawWireframe(MeshPart* part)
 
     case Mesh::TRIANGLE_STRIP:
         {
-            for (unsigned int i = 2; i < indexCount; ++i)
+            for (size_t i = 2; i < indexCount; ++i)
             {
                 GL_ASSERT( glDrawElements(GL_LINE_LOOP, 3, part->getIndexFormat(), ((const GLvoid*)((i-2)*indexSize))) );
             }
